hwc2: add locks for callback functions
In case surfaceflinger is killed, the function pointer of
RegisterCallback will be null. If the system is still handling
callback function tasks, it could have the chance to access null
pointer and result in a crash. This change adds locks for
callback functions to ensure the tasks are finished then we can
assign null pointer via RegisterCallback.

Bug: 150223963
Test: no crash while surfaceflinger is killed
Change-Id: I9ae8ac8968f4ff9286846de15df0411b5d79ae63
Signed-off-by: DennySPb <dennyspb@gmail.com>